Bug fixes (omero2): ------------------ * Got JDK5 code out of the client * Removed password field from Experimenter * Rearranged exception handling (not yet functional) * Fixed various dependency issues (oscache/mysql) * Fixed eclipse projects for new version
MISC:
----
* Added proto exception hierarcy



git-svn-id: file:///home/svn/omero/branches/omero2/OME_2_5@220 05709c45-44f0-0310-885b-81a1db45b4a6

+++ components/shoola-adapter/src/pojos/PixelsData.java
@@ -191,10 +191,10 @@ public class PixelsData
 			}
 			this.setPixelType(Model2PojosMapper.getPixelTypeID(pix.getPixelType()));
 			this.setSizeC(mapper.nullSafeInt(pix.getSizeC()));
-			this.setSizeT(mapper.nullSafeInt(pix.getSizeT().intValue()));
-			this.setSizeX(mapper.nullSafeInt(pix.getSizeX().intValue()));
-			this.setSizeY(mapper.nullSafeInt(pix.getSizeY().intValue()));
-			this.setSizeZ(mapper.nullSafeInt(pix.getSizeZ().intValue())); // TODO null error here?
+			this.setSizeT(mapper.nullSafeInt(pix.getSizeT()));
+			this.setSizeX(mapper.nullSafeInt(pix.getSizeX()));
+			this.setSizeY(mapper.nullSafeInt(pix.getSizeY()));
+			this.setSizeZ(mapper.nullSafeInt(pix.getSizeZ())); // TODO null error here?
 			
 		} else {
 			throw new IllegalArgumentException("PixelData copies only from ImagePixel");
